% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm_download.R
\name{gpm_download}
\alias{gpm_download}
\title{Download IMERG v06B Final, Early and Late Run}
\source{
https://github.com/csaybar/gpm/
}
\usage{
gpm_download(
  path,
  user,
  password,
  dates,
  product,
  band = "precipitationCal",
  lonMin = -86,
  lonMax = -66,
  latMin = -19.25,
  latMax = 1.25,
  removeHDF5 = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{path}{path (character). The default corresponds to the working directory, getwd(). Missing values will be ignored.}

\item{user}{pmm user (character). See https://pmm.nasa.gov/data-access/downloads/gpm}

\item{password}{pmm password (character). See https://pmm.nasa.gov/data-access/downloads/gpm}

\item{dates}{sequences of days.}

\item{product}{IMERG product (character). Currently are supported: 'finalrun', 'late' and 'early'.}

\item{band}{The available bands (character) are: \cr
band "HQobservationTime"  \cr
band "HQprecipSource"  \cr
band "HQprecipitation"  \cr
band "IRkalmanFilterWeight" \cr
band "IRprecipitation"\cr
band "precipitationCal"  \cr
band "precipitationUncal"  \cr
band "probabilityLiquidPrecipitation"  \cr
band "randomError"  \cr}

\item{lonMin}{Minimum latitude of bounding box}

\item{lonMax}{Maximum latitude of bounding box}

\item{latMin}{Minimum longitude of bounding box}

\item{latMax}{Maximum longitude of bounding box}

\item{removeHDF5}{Remove HD5 files after the conversion to NetCDF}

\item{quiet}{logical; suppress info of the download.}
}
\description{
Download IMERG v06B Final, Early and Late Run.
}
\note{
Modified by Ahmed Homoudi, November 2022.
}
\examples{
\dontrun{
gpm_download(
  path = ".",
  user = "you_user@gmail.com",
  password = "you_user@gmail.com",
  dates = c("2018-01-01"),
  product = "finalrun"
)
}
}
\author{
Cesar Aybar
}
