% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idps.R
\docType{package}
\name{idps}
\alias{idps}
\title{idps: A package for computating the notorious bar statistic.}
\description{
The idps package provides three categories of important functions:
foo, bar and baz.
}
\section{idps functions}{

The idps functions ...
}

