% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_netCDF_files.R
\name{separate_netCDF_files}
\alias{separate_netCDF_files}
\title{Aggregate netCDF files to monthly files.}
\usage{
separate_netCDF_files(netCDF.files, threshold, output.path = "~")
}
\arguments{
\item{netCDF.files}{A character vector listing all netCDF files}

\item{threshold}{A numeric value representing the minimum precipitation rate
that will be considered to delineate the precipitation objects, e.g. 5 mm/hr}

\item{output.path}{A character indicating where to place output files.}
}
\value{
netCDF file
}
\description{
A function to split the monthly netCDF files to precipitation
systems files, the splitting will search for time steps with precipitation less
than the provided threshold. Afterwards, this time step will be remove and
data after and before will be split to tmp netCDF files.
}
\author{
Ahmed Homoudi
}
